extract save button code to separate file
