+++ src/js/main.js
@@ -10,7 +10,8 @@ import {
   selectAllCategories,
   toggleSelectAllCheckbox
 } from "./vol/filters-form";
-import { updateSaveButton, renderSavedOpps } from "./vol/dom-saved-opps";
+import { renderSavedOpps } from "./vol/dom-saved-opps";
+import { updateSaveButton } from "./vol/save-button";
 
 // Webpack uses this file as entry point for bundling CSS assets
 import "../css/index.js";
