+++ src/js/vol/dom-saved-opps.js
@@ -39,20 +39,6 @@ function showNoOppsMessage(parentElt) {
     '<h2>Nothing Saved Yet</h2><p><a href="/opportunities">Continue browsing opportunities</a></p>';
 }
 
-// change the appearance and text of the save button when an opp is saved or removed
-export function updateSaveButton(isSaved) {
-  if (isSaved) {
-    this.classList.remove("btn-default");
-    this.classList.add("btn-success");
-    this.innerHTML = 'Saved!&nbsp;<span class="glyphicon glyphicon-ok"></span>';
-  } else {
-    this.classList.add("btn-default");
-    this.classList.remove("btn-success");
-    this.innerHTML =
-      'Save&nbsp;<span class="glyphicon glyphicon-pushpin"></span>';
-  }
-}
-
 // Update the display
 export function renderSavedOpps(store, parentElt) {
   // Remove everything currently in parentElt
